% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotplayer_funct.R
\name{plotplayer}
\alias{plotplayer}
\title{create a barplot with values}
\usage{
plotplayer(p)
}
\arguments{
\item{id}{character. the name of the player you're interested in}
}
\value{
barplot
x axis default season
y axis playername selected as param
}
\description{
this function enables to visualize graphically the performace of a single player in the season considered in the dataset
}
\examples{
plotplayer("playername")

}
