% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalscore_funct.R
\name{totalscore}
\alias{totalscore}
\title{sum the values by column}
\usage{
totalscore(dataset = NULL, p = "BELOTTI")
}
\arguments{
\item{id}{character}
}
\value{
integer
}
\description{
this function enables to sum the score of a player in all the considered season and obtain the total score
}
\examples{
totalscore("playername")

}
